refactor: move SlogToLogBridge to separate file
